LPS-108755 Move classes and rename
